requirements.txt
flask==2.0.1
flask-login==0.5.0
pymongo==3.12.0
flask-pymongo==2.3.0
python-dotenv==0.19.0
werkzeug==2.0.1
bcrypt==3.2.0